Admin Middleware

+++ routes/web.php
@@ -20,9 +20,13 @@ Route::post('/create', [SessionCtrl::class, 'create']);
 Route::get('/login', [SessionCtrl::class, 'indexLogin']);
 Route::get('/logout', [SessionCtrl::class, 'logout'])->name('logout');
 
-Route::get('/dashboard', [SessionCtrl::class, 'dashboard']);
-Route::post('/dashboard', [SessionCtrl::class, 'login']);
-Route::get('/dashboard', [SessionCtrl::class, 'showData']); // show database data
-Route::get('/editData/{id}', [SessionCtrl::class, 'edit'])->name('editData');
-Route::delete('/delete/{id}', [SessionCtrl::class, 'delete'])->name('delete');
-Route::patch('/updateData/{id}', [SessionCtrl::class, 'update'])->name('update');
+
+Route::get('/dashboard', [SessionCtrl::class, 'dashboard'])->middleware('isAdmin');
+Route::post('/dashboard', [SessionCtrl::class, 'login'])->middleware('isAdmin');
+Route::get('/dashboard', [SessionCtrl::class, 'showData'])->middleware('isAdmin'); // show database data
+Route::get('/editData/{id}', [SessionCtrl::class, 'edit'])->name('editData')->middleware('isAdmin');
+Route::delete('/delete/{id}', [SessionCtrl::class, 'delete'])->name('delete')->middleware('isAdmin');
+Route::patch('/updateData/{id}', [SessionCtrl::class, 'update'])->name('update')->middleware('isAdmin');
+
+Route::get('/user', [SessionCtrl::class, 'indexUser']);
+Route::post('/user', [SessionCtrl::class, 'login']);
